server/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Cors==3.0.10
python-dotenv==1.0.0

# ML / Computer Vision
opencv-python-headless==4.7.0.72
dlib==19.24.2
face-recognition==1.3.0

# For image processing
Pillow==9.5.0

# Optional: for testing
pytest==7.3.1
pytest-flask==1.2.0
